+++ modules/renderer/src/com/genir/renderer/bridge/commands/GL15.java
@@ -55,6 +55,8 @@ public class GL15 {
         record glBindBuffer(int target, int buffer) implements GLCommand {
             @Override
             public void run(Context context, float[] args, int argsOffset) {
+                context.attribManager.glBindBuffer(target, buffer);
+
                 org.lwjgl.opengl.GL15.glBindBuffer(target, buffer);
             }
         }