+++ modules/renderer/src/com/genir/renderer/bridge/context/AttribManager.java
@@ -3,6 +3,7 @@ package com.genir.renderer.bridge.context;
 import com.genir.renderer.bridge.context.stall.AttribState;
 import org.lwjgl.opengl.GL11;
 import org.lwjgl.opengl.GL13;
+import org.lwjgl.opengl.GL15;
 
 import java.util.Map;
 import java.util.Stack;
@@ -95,6 +96,25 @@ public class AttribManager {
         }
     }
 
+    // Apply array buffer binding selected by the client.
+    public void applyArrayBufferBinding() {
+        if (actual.arrayBufferBinding != expected.arrayBufferBinding) {
+            actual.arrayBufferBinding = expected.arrayBufferBinding;
+
+            org.lwjgl.opengl.GL15.glBindBuffer(GL15.GL_ARRAY_BUFFER, expected.arrayBufferBinding);
+        }
+    }
+
+    // Set array buffer binding required by the bridge, which may be
+    // different from binding selected by the client.
+    public void forceArrayBufferBinding(int buffer) {
+        if (actual.arrayBufferBinding != buffer) {
+            actual.arrayBufferBinding = buffer;
+
+            org.lwjgl.opengl.GL15.glBindBuffer(GL15.GL_ARRAY_BUFFER, buffer);
+        }
+    }
+
     //
     // GL calls.
     //
@@ -135,17 +155,21 @@ public class AttribManager {
     }
 
     public void glBindTexture(int target, int texture) {
+        // Texture binding is not permanently overriden.
         expected.glBindTexture(target, texture);
-
-        // Texture binding is not overriden, just tracked.
         actual.glBindTexture(target, texture);
     }
 
     public void glActiveTexture(int mode) {
+        // Texture unit is not permanently overriden.
         expected.glActiveTexture(mode);
-
-        // Texture unit is not overriden, just tracked.
         actual.glActiveTexture(mode);
+    }
+
+    public void glBindBuffer(int target, int buffer) {
+        // Buffer binding is not permanently overriden.
+        expected.glBindBuffer(target, buffer);
+        actual.glBindBuffer(target, buffer);
     }
 
     public void glBlendFunc(int sfactorRGB, int dfactorRGB) {